@@ -1,53 +1,43 @@
-<template>
-  <div>
-    <nuxt/>
-  </div>
+<template lang="pug">
+  v-app(dark)
+    v-toolbar(app)
+      v-toolbar-title {{ title }}
+      v-spacer
+      v-toolbar-items
+        v-layout(align-center)
+          v-flex
+            v-chip(disabled)
+              v-avatar.blue
+                v-icon border_horizontal
+              | {{ latitude }}
+            v-chip(disabled)
+              v-avatar.blue
+                v-icon border_vertical
+              | {{ longitude }}
+    v-content
+      nuxt
+    v-footer(app)
+      v-layout(align-center, justify-center)
+        v-btn
+          v-icon place
+        v-btn
+          v-icon directions_walk
+        v-btn
+          v-icon directions_run
+        v-btn
+          v-icon local_taxi
 </template>
 
-<style>
-html {
-  font-family: "Source Sans Pro", -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif;
-  font-size: 16px;
-  word-spacing: 1px;
-  -ms-text-size-adjust: 100%;
-  -webkit-text-size-adjust: 100%;
-  -moz-osx-font-smoothing: grayscale;
-  -webkit-font-smoothing: antialiased;
-  box-sizing: border-box;
-}
+<script>
+import configs from '../configs'
 
-*, *:before, *:after {
-  box-sizing: border-box;
-  margin: 0;
+export default {
+  data() {
+    return {
+      title: configs.title,
+      latitude: '-',
+      longitude: '-'
+    }
+  }
 }
-
-.button--green {
-  display: inline-block;
-  border-radius: 4px;
-  border: 1px solid #3b8070;
-  color: #3b8070;
-  text-decoration: none;
-  padding: 10px 30px;
-}
-
-.button--green:hover {
-  color: #fff;
-  background-color: #3b8070;
-}
-
-.button--grey {
-  display: inline-block;
-  border-radius: 4px;
-  border: 1px solid #35495e;
-  color: #35495e;
-  text-decoration: none;
-  padding: 10px 30px;
-  margin-left: 15px;
-}
-
-.button--grey:hover {
-  color: #fff;
-  background-color: #35495e;
-}
-</style>
-
+</script>
